pyproject.toml
[tool.poetry]
name = "orderline_dw"
version = "0.1.0"
description = ""
authors = ["Mischa Barthel <barthel.m@hsleiden.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
prefect = "^2.16.0"
click = "^8.1.7"
python-dotenv = "^1.0.1"
pandas = "^2.2.0"
sqlalchemy = "^2.0.25"
pyarrow = "^15.0.0"
pyodbc = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
prefect-flows = "orderline_dw.cli:main"
